+++ CHANGELOG.md
@@ -13,16 +13,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Changed Mixins to only attempt to load if the appropriate minimap mod is loaded
 
 ### Fixed
-- Fixed icon alignment if XaeroMinimap's size is chunk coordinate size is wider than the minimap size
+- Fixed icon alignment if Xaero's Minimap size is chunk coordinate size is wider than the minimap size
+- Fixed icon alignment when using the Enlarged Minimap feature in Xaero's Minimap
 - Fixed icon alignment if using OpenPartiesAndClaims, and the claim name splits into more than one line.
-- Generally improved XaeroMinimap's season icon placement
 
 
 ## [1.5.13] - 2023-11-01
 
 ### Fixed
 - Adjusted math behind icon position
-- (Xaeros Minimap) Fixed icon adjusting when hidden potion effects are active
+- (Xaero's Minimap) Fixed icon adjusting when hidden potion effects are active
 
 
 ## [1.5.12] - 2023-10-22